+++ R/Design-methods.R
@@ -257,13 +257,13 @@ setMethod("simulate",
                       thisProb <- thisTruth(thisDose)
 
                       ## what is the cohort size at this dose?
-                      thisSize <- size(cohortSize=object@cohortSize,
+                      thisSize <- size(object@cohortSize,
                                        dose=thisDose,
                                        data=thisData)
 
                       ## In case there are placebo
                       if(thisData@placebo)
-                          thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                          thisSize.PL <- size(object@PLcohortSize,
                                               dose=thisDose,
                                               data=thisData)
 
@@ -505,7 +505,7 @@ setMethod("simulate",
                       thisProb <- thisTruth(thisDose)
 
                       ## what is the cohort size at this dose?
-                      thisSize <- size(cohortSize=object@cohortSize,
+                      thisSize <- size(object@cohortSize,
                                        dose=thisDose,
                                        data=thisData)
 
@@ -712,14 +712,14 @@ setMethod("simulate",
                       thisMeanBiomarker <- thisTrueBiomarker(thisDose)
 
                       ## what is the cohort size at this dose?
-                      thisSize <- size(cohortSize=object@cohortSize,
+                      thisSize <- size(object@cohortSize,
                                        dose=thisDose,
                                        data=thisData)
 
                       ## In case there are placebo
                       ## what is the cohort size at this dose for Placebo?
                       if(thisData@placebo)
-                          thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                          thisSize.PL <- size(object@PLcohortSize,
                                               dose=thisDose,
                                               data=thisData)
 
@@ -1013,12 +1013,12 @@ setMethod("examine",
                   while(! stopit)
                   {
                       ## what is the cohort size at this dose?
-                      thisSize <- size(cohortSize=object@cohortSize,
+                      thisSize <- size(object@cohortSize,
                                        dose=thisDose,
                                        data=baseData)
 
                       if(baseData@placebo)
-                        thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                        thisSize.PL <- size(object@PLcohortSize,
                                             dose=thisDose,
                                             data=baseData)
 
@@ -1190,7 +1190,7 @@ setMethod("examine",
                   while(! stopit)
                   {
                       ## what is the cohort size at this dose?
-                      thisSize <- size(cohortSize=object@cohortSize,
+                      thisSize <- size(object@cohortSize,
                                        dose=thisDose,
                                        data=baseData)
 
@@ -1373,7 +1373,7 @@ setMethod("examine",
               {
 
                 ## what is the cohort size at this dose?
-                thisSize <- size(cohortSize=object@cohortSize,
+                thisSize <- size(object@cohortSize,
                                  dose=thisDose,
                                  data=baseData)
 
@@ -1734,13 +1734,13 @@ setMethod("simulate",
 
 
                   ## what is the cohort size at this dose?
-                  thisSize <- size(cohortSize=object@cohortSize,
+                  thisSize <- size(object@cohortSize,
                                    dose=thisDose,
                                    data=thisData)
 
                   ## In case there are placebo
                   if(thisData@placebo)
-                    thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                    thisSize.PL <- size(object@PLcohortSize,
                                         dose=thisDose,
                                         data=thisData)
 
@@ -2046,13 +2046,13 @@ setMethod("simulate",
                   thisProb <- thisTruth(thisDose)
 
                   ## what is the cohort size at this dose?
-                  thisSize <- size(cohortSize=object@cohortSize,
+                  thisSize <- size(object@cohortSize,
                                    dose=thisDose,
                                    data=thisData)
 
                   ## In case there are placebo
                   if(thisData@placebo)
-                    thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                    thisSize.PL <- size(object@PLcohortSize,
                                         dose=thisDose,
                                         data=thisData)
 
@@ -2384,7 +2384,7 @@ setMethod("simulate",
                   thisMeanEff<-thisTruthEff(thisDose)
 
                   ## what is the cohort size at this dose?
-                  thisSize <- size(cohortSize=object@cohortSize,
+                  thisSize <- size(object@cohortSize,
                                    dose=thisDose,
                                    data=thisData)
 
@@ -2392,7 +2392,7 @@ setMethod("simulate",
                   ## In case there are placebo
                   ## what is the cohort size at this dose for Placebo?
                   if(thisData@placebo)
-                    thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                    thisSize.PL <- size(object@PLcohortSize,
                                         dose=thisDose,
                                         data=thisData)
 
@@ -2856,12 +2856,12 @@ setMethod("simulate",
 
 
                     ## what is the cohort size at this dose?
-                    thisSize <- size(cohortSize=object@cohortSize,
+                    thisSize <- size(object@cohortSize,
                                      dose=thisDose,
                                      data=thisData)
 
                     if(thisData@placebo)
-                      thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                      thisSize.PL <- size(object@PLcohortSize,
                                           dose=thisDose,
                                           data=thisData)
 
@@ -3260,7 +3260,7 @@ setMethod("simulate",
                   thisMeanEff<-thisTruthEff(thisDose)
 
                   ## what is the cohort size at this dose?
-                  thisSize <- size(cohortSize=object@cohortSize,
+                  thisSize <- size(object@cohortSize,
                                    dose=thisDose,
                                    data=thisData)
 
@@ -3798,7 +3798,7 @@ setMethod("simulate",
                   thisProb <- thisTruth(thisDose)
 
                   ## what is the cohort size at this dose?
-                  thisSize <- size(cohortSize=object@cohortSize,
+                  thisSize <- size(object@cohortSize,
                                    dose=thisDose,
                                    data=thisData)
 
@@ -3807,7 +3807,7 @@ setMethod("simulate",
 
                   ## In case there are placebo
                   if(thisData@placebo)
-                    thisSize.PL <- size(cohortSize=object@PLcohortSize,
+                    thisSize.PL <- size(object@PLcohortSize,
                                         dose=thisDose,
                                         data=thisData)